allow uin32_t bauds https://github.com/adafruit/Adafruit_GPS/issues/42

+++ Adafruit_GPS.cpp
@@ -350,7 +350,7 @@ void Adafruit_GPS::common_init(void) {
     speed = angle = magvariation = HDOP = 0.0; // float
 }
 
-void Adafruit_GPS::begin(uint16_t baud)
+void Adafruit_GPS::begin(uint32_t baud)
 {
 #ifdef __AVR__
   if(gpsSwSerial)
